+++ src/figure_agent/fsl/__init__.py
@@ -1,0 +1,44 @@
+"""Figure Specification Language (FSL) engine."""
+
+from figure_agent.fsl.exceptions import (
+    FSLError,
+    FSLParseError,
+    FSLSchemaError,
+    FSLValidationError,
+)
+from figure_agent.fsl.models import (
+    ExportOptions,
+    Figure,
+    Layout,
+    Metadata,
+    ObjectReference,
+    Panel,
+    StyleReference,
+    ValidationOptions,
+)
+from figure_agent.fsl.parser import load_json, load_yaml, parse, validate_schema
+from figure_agent.fsl.serializer import to_dict, to_json, to_yaml
+from figure_agent.fsl.validator import FSLValidator
+
+__all__ = [
+    "ExportOptions",
+    "FSLError",
+    "FSLParseError",
+    "FSLSchemaError",
+    "FSLValidationError",
+    "FSLValidator",
+    "Figure",
+    "Layout",
+    "Metadata",
+    "ObjectReference",
+    "Panel",
+    "StyleReference",
+    "ValidationOptions",
+    "load_json",
+    "load_yaml",
+    "parse",
+    "to_dict",
+    "to_json",
+    "to_yaml",
+    "validate_schema",
+]